+++ authobj.h
@@ -35,18 +35,18 @@ struct _auth_obj {
 	unsigned char *buffer;	/* to be free()'d if not NULL */
 	const char *err;	/* non-NULL if failed */
 	unsigned char *data;
-	int datasize;
+	size_t datasize;
 	unsigned char *payload;
-	int paylsize;
+	size_t paylsize;
 };
 
 /* Construct new or repack old authobj, return payload */
 struct _auth_obj authobj(const char *userid, const char *password,
 		const char *oldnonce, const char *newnonce,
-		const unsigned char *secret, const int secsize,
-		const unsigned char *payload, const int paysize,
-		const unsigned char *ablob, const int blobsize,
+		const unsigned char *secret, const size_t secsize,
+		const unsigned char *payload, const size_t paysize,
+		const unsigned char *ablob, const size_t blobsize,
 		struct _auth_chunk (*fetch_key)(const unsigned char *chal,
-						const int csize));
+						const size_t csize));
 
 #endif
