+++ crypto_if.h
@@ -26,15 +26,15 @@ freely, subject to the following restrictions:
 
 struct crypto_interface {
 	const char *(*init)(void);
-	unsigned long (*encrypt)(const void *key, const int keylen, void *iv,
-				const void *pt, void *ct, const int tlen);
-	unsigned long (*decrypt)(const void *key, const int keylen, void *iv,
-				const void *ct, void *pt, const int tlen);
-	unsigned long (*hash)(const void *pt, const int tlen,
-				void *tag, int *taglen);
-	unsigned long (*hmac)(const void *key, const int keylen,
-				const void *pt, const int tlen,
-				void *tag, int *taglen);
+	unsigned long (*encrypt)(const void *key, const size_t keylen, void *iv,
+				const void *pt, void *ct, const size_t tlen);
+	unsigned long (*decrypt)(const void *key, const size_t keylen, void *iv,
+				const void *ct, void *pt, const size_t tlen);
+	unsigned long (*hash)(const void *pt, const size_t tlen,
+				void *tag, size_t *taglen);
+	unsigned long (*hmac)(const void *key, const size_t keylen,
+				const void *pt, const size_t tlen,
+				void *tag, size_t *taglen);
 	const char *(*errstr)(const unsigned long err);
 };
