+++ gnu_crypto.c
@@ -36,8 +36,8 @@ static const char *gnu_init(void)
 	return "gcrypt";
 }
 
-static unsigned long gnu_encrypt(const void *key, const int keylen, void *iv,
-			const void *pt, void *ct, const int tlen)
+static unsigned long gnu_encrypt(const void *key, const size_t keylen, void *iv,
+			const void *pt, void *ct, const size_t tlen)
 {
 	gcry_error_t err;
 	gcry_cipher_hd_t hd;
@@ -56,8 +56,8 @@ static unsigned long gnu_encrypt(const void *key, const int keylen, void *iv,
 	return 0UL;
 }
 
-static unsigned long gnu_decrypt(const void *key, const int keylen, void *iv,
-			const void *ct, void *pt, const int tlen)
+static unsigned long gnu_decrypt(const void *key, const size_t keylen, void *iv,
+			const void *ct, void *pt, const size_t tlen)
 {
 	gcry_error_t err;
 	gcry_cipher_hd_t hd;
@@ -76,8 +76,8 @@ static unsigned long gnu_decrypt(const void *key, const int keylen, void *iv,
 	return 0UL;
 }
 
-static unsigned long gnu_hash(const void *pt, const int tlen,
-			void *tag, int *taglen)
+static unsigned long gnu_hash(const void *pt, const size_t tlen,
+			void *tag, size_t *taglen)
 {
 	gcry_error_t err;
 	gcry_md_hd_t hd;
@@ -95,9 +95,9 @@ static unsigned long gnu_hash(const void *pt, const int tlen,
 	return 0UL;
 }
 
-static unsigned long gnu_hmac(const void *key, const int keylen,
-			const void *pt, const int tlen,
-			void *tag, int *taglen)
+static unsigned long gnu_hmac(const void *key, const size_t keylen,
+			const void *pt, const size_t tlen,
+			void *tag, size_t *taglen)
 {
 	gcry_error_t err;
 	gcry_md_hd_t hd;
