+++ pcsc_cr.h
@@ -25,8 +25,8 @@ freely, subject to the following restrictions:
 #define _PCSC_CR_H
 
 int pcsc_option(const char *option);
-long pcsc_cr(const unsigned char *chal, const int csize,
-		unsigned char *resp, int *rsize);
+long pcsc_cr(const unsigned char *chal, const size_t csize,
+		unsigned char *resp, size_t *rsize);
 char *pcsc_errstr(long err);
 
 #endif
