+++ tom_crypto.c
@@ -34,8 +34,8 @@ static const char *tom_init(void)
 	return "tomcrypt";
 }
 
-static unsigned long tom_encrypt(const void *key, const int keylen, void *iv,
-			const void *pt, void *ct, const int tlen)
+static unsigned long tom_encrypt(const void *key, const size_t keylen, void *iv,
+			const void *pt, void *ct, const size_t tlen)
 {
 	symmetric_CBC cbc;
 	int index, err;
@@ -49,8 +49,8 @@ static unsigned long tom_encrypt(const void *key, const int keylen, void *iv,
 	return err;
 }
 
-static unsigned long tom_decrypt(const void *key, const int keylen, void *iv,
-			const void *ct, void *pt, const int tlen)
+static unsigned long tom_decrypt(const void *key, const size_t keylen, void *iv,
+			const void *ct, void *pt, const size_t tlen)
 {
 	symmetric_CBC cbc;
 	int index, err;
@@ -64,8 +64,8 @@ static unsigned long tom_decrypt(const void *key, const int keylen, void *iv,
 	return err;
 }
 
-static unsigned long tom_hash(const void *pt, const int tlen,
-			void *tag, int *taglen)
+static unsigned long tom_hash(const void *pt, const size_t tlen,
+			void *tag, size_t *taglen)
 {
 	int index, rc;
 	unsigned long ltaglen = *taglen;
@@ -77,9 +77,9 @@ static unsigned long tom_hash(const void *pt, const int tlen,
 	return rc;
 }
 
-static unsigned long tom_hmac(const void *key, const int keylen,
-			const void *pt, const int tlen,
-			void *tag, int *taglen)
+static unsigned long tom_hmac(const void *key, const size_t keylen,
+			const void *pt, const size_t tlen,
+			void *tag, size_t *taglen)
 {
 	int index, rc;
 	unsigned long ltaglen = *taglen;
